Missende constructors toegevoegd

+++ codecademy2/src/main/java/com/codecademy2/Domain/Certificate.java
@@ -3,15 +3,17 @@ package com.codecademy2.Domain;
 public class Certificate{
 
 private String certificateID;
-private int grade;
+private double grade;
 private String employee;
 
-    public Certificate(String certificateID, int grade, String employee) {
+public Certificate(String certificateID, double grade, String employee) {
         this.certificateID = certificateID;
         this.grade = grade;
         this.employee = employee;
     }
 
+
+    
     public String getCertificateID() {
         return certificateID;
     }
